[feat] 認証する/しないの設定
