+++ src/main.ts
@@ -10,15 +10,40 @@ const guildRouter = require("./routes/guild");
 const channelRouter = require("./routes/channel");
 const messageRouter = require("./routes/message");
 const mediaRouter = require("./routes/media");
-const jwttest = require("./routes/jwt_test");
+
+import jwt from "express-jwt";
+import jwksRsa from "jwks-rsa";
+import dotenv from "dotenv";
+
+const conf = dotenv.config();
+
+if (!process.env.AUTH0_DOMAIN || !process.env.AUTH0_AUDIENCE) {
+  throw 'Make sure you have AUTH0_DOMAIN, and AUTH0_AUDIENCE in your .env file';
+}
+
+export const checkJwt = jwt({
+  // Dynamically provide a signing key based on the [Key ID](https://tools.ietf.org/html/rfc7515#section-4.1.4) header parameter ("kid") and the signing keys provided by the JWKS endpoint.
+  secret: jwksRsa.expressJwtSecret({
+    cache: true,
+    rateLimit: true,
+    jwksRequestsPerMinute: 5,
+    jwksUri: `https://${process.env.AUTH0_DOMAIN}/.well-known/jwks.json`
+  }),
+
+  // Validate the audience and the issuer.
+  audience: process.env.AUTH0_AUDIENCE,
+  issuer: [`https://${process.env.AUTH0_DOMAIN}/`],
+  algorithms: ['RS256']
+});
+
+
 
 app.use("/", infoRouter);
-app.use("/users", userRouter);
-app.use("/guilds", guildRouter);
-app.use("/", channelRouter);
-app.use("/channels", messageRouter);
-app.use("/files", mediaRouter);
-app.use("/jwt", jwttest);
+app.use("/users", checkJwt, userRouter);
+app.use("/guilds", checkJwt, guildRouter);
+app.use("/", checkJwt, channelRouter);
+app.use("/channels", checkJwt, messageRouter);
+app.use("/files", checkJwt, mediaRouter);
 
 app.listen(3080,() => {
   const black   = '\u001b[30m';
